# openapi.yaml
openapi: 3.0.3
info:
  title: Threshwork
  version: 0.1.0
  description: >-
    Automation portfolio control plane for content operations. Scores every review
    queue on capacity, difficulty, and impact; sets the confidence threshold at which
    a machine decision stands without a human; routes work between classifier vendors
    and human agents under statutory clocks and rights gates; and meters realised
    benefit per queue against the business case that funded it.
  x-product-slug: ai-content-threshwork
  x-source: ai-in-content/Accenture-AI-In-Content-Services/
servers:
  - url: https://api.example.com
...
paths:
  /v1/queues:
    get:
      operationId: listQueues
      tags: [Queues]
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: impactTier
          in: query
          schema:
            $ref: "#/components/schemas/ImpactTier"
        - name: automationStatus
          in: query
          schema:
            type: string
            enum: [not_assessed, blocked, approved_pilot, approved_scaled, suspended, reverted]
      responses:
        "200":
          description: Queues page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeQueue"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: registerQueue
      tags: [Queues]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueueCreate"
      responses:
        "201":
          description: Queue registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeQueue"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /v1/queues/{queueId}:
    parameters:
      - $ref: "#/components/parameters/QueueId"
    get:
      operationId: getQueue
      tags: [Queues]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeQueue"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/queues/{queueId}/scorecard:
    parameters:
      - $ref: "#/components/parameters/QueueId"
    get:
      operationId: getQueueScorecard
      tags: [Scorecards]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current weighted scorecard for the queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeQueueScorecard"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: scoreQueue
      tags: [Scorecards]
      description: >-
        Run a scorecard against agreed weightings. Weightings must be supplied and
        are recorded with the score so the ranking is reproducible later.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueueScorecardCreate"
      responses:
        "201":
          description: Scorecard recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeQueueScorecard"
...
  /v1/appeals/{appealId}/resolution:
    parameters:
      - $ref: "#/components/parameters/AppealId"
    post:
      operationId: resolveAppeal
      tags: [Appeals]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [outcome, rationale]
              properties:
                outcome:
                  type: string
                  enum: [upheld, rejected]
                rationale:
                  type: string
                feedsCalibrationSet:
                  type: boolean
                  default: true
      responses:
        "200":
          description: Appeal resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeAppeal"
        "403":
          description: Reviewer made the original decision and may not adjudicate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /v1/vendors/{vendorId}/evaluations:
    parameters:
      - $ref: "#/components/parameters/VendorId"
    get:
      operationId: listVendorEvaluations
      tags: [Vendors]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Evaluations page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeVendorEvaluation"
    post:
      operationId: startVendorEvaluation
      tags: [Vendors]
      description: >-
        Run a candidate vendor in shadow against the operation's own golden set and
        optionally against live queue traffic, without acting on the result.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorEvaluationCreate"
      responses:
        "202":
          description: Shadow evaluation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeVendorEvaluation"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    QueueId:
      name: queueId
      in: path
      required: true
      schema:
        type: string
...
    AppealId:
      name: appealId
      in: path
      required: true
      schema:
        type: string
    VendorId:
      name: vendorId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    ResponseMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
          format: double
        currency:
          type: string
          example: USD
    ImpactTier:
      type: string
      description: >-
        Consequence of a wrong decision in this queue. Illegal content admitted to a
        platform and a mis-tagged product image do not share a threshold.
      enum: [critical_illegal, high_safety_policy, moderate_brand_or_accessibility, low_enrichment]
    TaskType:
      type: string
      enum:
        - policy_moderation
        - offensive_media_review
        - metadata_enrichment
        - catalogue_tagging
        - localisation_qa
        - transcreation_review
        - accessibility_captioning
        - accessibility_alt_text
        - rights_clearance
        - editorial_approval
...
    Queue:
      type: object
      required: [id, name, taskType, impactTier, automationStatus]
      properties:
        id:
          type: string
        name:
          type: string
        taskType:
          $ref: "#/components/schemas/TaskType"
        impactTier:
          $ref: "#/components/schemas/ImpactTier"
        automationStatus:
          type: string
          enum: [not_assessed, blocked, approved_pilot, approved_scaled, suspended, reverted]
        deploymentMode:
          type: string
          description: Deployment choice from the source WHY-WHERE-HOW framework
          enum: [human_only, human_in_the_loop, machine_final_with_sampling, machine_final]
        activeVendorId:
          type: string
        fallbackVendorId:
          type: string
        policyDefinitionId:
          type: string
        dailyVolume:
          type: integer
        averageHandlingTimeSeconds:
          type: integer
        qualityThreshold:
          type: number
          format: float
        territoryRestrictions:
          type: array
          items:
            type: string
    QueueCreate:
      type: object
      required: [name, taskType, impactTier]
      properties:
        name:
          type: string
        taskType:
          $ref: "#/components/schemas/TaskType"
        impactTier:
          $ref: "#/components/schemas/ImpactTier"
        policyDefinitionId:
          type: string
        dailyVolume:
          type: integer
        averageHandlingTimeSeconds:
          type: integer
        qualityThreshold:
          type: number
          format: float
        territoryRestrictions:
          type: array
          items:
            type: string
...
    QueueScorecard:
      type: object
      required: [id, queueId, weightedScore, scoredAt]
      properties:
        id:
          type: string
        queueId:
          type: string
        weightedScore:
          type: number
          format: float
        rank:
          type: integer
        capacityScore:
          type: number
          format: float
          description: Current queue load against handling time and quality targets
        anticipatedDemandScore:
          type: number
          format: float
          description: Forecast volume the queue cannot absorb with current headcount
        humanDifficultyScore:
          type: number
          format: float
          description: How hard or time-consuming the task is for an agent
        automationDifficultyScore:
          type: number
          format: float
          description: How hard the task is for off-the-shelf classifiers
        impactScore:
          type: number
          format: float
        weightings:
          type: object
          additionalProperties:
            type: number
            format: float
        scoredBy:
          type: string
        scoredAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    QueueScorecardCreate:
      type: object
      required: [weightings, capacityScore, anticipatedDemandScore, humanDifficultyScore, automationDifficultyScore, impactScore]
      properties:
        weightings:
          type: object
          additionalProperties:
            type: number
            format: float
        capacityScore:
          type: number
          format: float
        anticipatedDemandScore:
          type: number
          format: float
        humanDifficultyScore:
          type: number
          format: float
        automationDifficultyScore:
          type: number
          format: float
        impactScore:
          type: number
          format: float
        notes:
          type: string
...
    VendorEvaluation:
      type: object
      required: [id, vendorId, queueId, mode, status]
      properties:
        id:
          type: string
        vendorId:
          type: string
        queueId:
          type: string
        goldenSetId:
          type: string
        mode:
          type: string
          enum: [golden_set_only, live_shadow, golden_set_and_live_shadow]
        status:
          type: string
          enum: [queued, running, complete, abandoned]
        agreementWithHumanLabel:
          type: number
          format: float
        falseNegativeRate:
          type: number
          format: float
        falsePositiveRate:
          type: number
          format: float
        residualHumanHandlingSeconds:
          type: integer
          description: Human time the vendor leaves behind per hundred items
        totalCostPerThousandDecisions:
          $ref: "#/components/schemas/Money"
        meetsContractedFloor:
          type: boolean
        completedAt:
          type: string
          format: date-time
    VendorEvaluationCreate:
      type: object
      required: [queueId, mode]
      properties:
        queueId:
          type: string
        goldenSetId:
          type: string
        mode:
          type: string
          enum: [golden_set_only, live_shadow, golden_set_and_live_shadow]
        sampleSize:
          type: integer
...
    Appeal:
      type: object
      required: [id, workItemId, status]
      properties:
        id:
          type: string
        workItemId:
          type: string
        raisedBy:
          type: string
          enum: [end_user, enterprise_client, trusted_flagger, internal]
        originalDecisionBy:
          type: string
          enum: [machine, human]
        originalReviewerId:
          type: string
        assignedReviewerId:
          type: string
        status:
          type: string
          enum: [pending, upheld, rejected, expired]
        dueAt:
          type: string
          format: date-time
        rationale:
          type: string
...
    DataEnvelopeQueue:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/Queue"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    DataEnvelopeQueueScorecard:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/QueueScorecard"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    DataEnvelopeAppeal:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/Appeal"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    DataEnvelopeVendorEvaluation:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/VendorEvaluation"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    ListEnvelopeQueue:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Queue"
            nextCursor:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    ListEnvelopeVendorEvaluation:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/VendorEvaluation"
            nextCursor:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...