# platform/api-server/src/domains/queues/handlers/queue.handlers.ts
/**
 * Handlers for Queue
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: queues
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { QueuesDomainModule } from "../dependencies/queues-ddd.dependencies.js";
import type { GetQueueInput, ListQueuesInput, RegisterQueueInput } from "@ddd/services/queues";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/queues
 * Operation: listQueues
 */
export async function listQueues(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], impactTier?: string | string[], automationStatus?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: QueuesDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListQueuesInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.queues.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/queues
 * Operation: registerQueue
 */
export async function registerQueue(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: QueuesDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: RegisterQueueInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.queues.create.execute(input);
  return reply.code(201).send(result);
}


/**
 * Handler for GET /v1/queues/{queueId}
 * Operation: getQueue
 */
export async function getQueue(
  request: FastifyRequest<{ Params: { queueId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: QueuesDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GetQueueInput = {
    queueId: request.params.queueId,
  };
  const result = await deps.useCases.queues.get.execute(input);
  return reply.code(200).send(result);
}