# platform/api-server/src/domains/queues/handlers/scorecard.handlers.ts
/**
 * Handlers for Scorecard
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: queues
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { QueuesDomainModule } from "../dependencies/queues-ddd.dependencies.js";
import type { GetQueueScorecardInput, ScoreQueueInput } from "@ddd/services/queues";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/queues/{queueId}/scorecard
 * Operation: getQueueScorecard
 */
export async function getQueueScorecard(
  request: FastifyRequest<{ Params: { queueId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: QueuesDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GetQueueScorecardInput = {
    queueId: request.params.queueId,
  };
  const result = await deps.useCases.scorecards.get.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/queues/{queueId}/scorecard
 * Operation: scoreQueue
 */
export async function scoreQueue(
  request: FastifyRequest<{ Params: { queueId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: QueuesDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ScoreQueueInput = {
    queueId: request.params.queueId,
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.scorecards.get.execute(input);
  return reply.code(201).send(result);
}