# platform/api-server/src/domains/routing/handlers/resolution.handlers.ts
/**
 * Handlers for Resolution
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: routing
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { RoutingDomainModule } from "../dependencies/routing-ddd.dependencies.js";
import type { ResolveAppealInput } from "@ddd/services/routing";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for POST /v1/appeals/{appealId}/resolution
 * Operation: resolveAppeal
 */
export async function resolveAppeal(
  request: FastifyRequest<{ Params: { appealId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: RoutingDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ResolveAppealInput = {
    appealId: request.params.appealId,
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.resolutions.get.execute(input);
  return reply.code(200).send(result);
}