# platform/api-server/src/domains/vendors/handlers/evaluation.handlers.ts
/**
 * Handlers for Evaluation
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: vendors
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { VendorsDomainModule } from "../dependencies/vendors-ddd.dependencies.js";
import type { ListVendorEvaluationsInput, StartVendorEvaluationInput } from "@ddd/services/vendors";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/vendors/{vendorId}/evaluations
 * Operation: listVendorEvaluations
 */
export async function listVendorEvaluations(
  request: FastifyRequest<{ Params: { vendorId: string }; Querystring: { cursor?: string | string[], limit?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: VendorsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListVendorEvaluationsInput = {
    vendorId: request.params.vendorId,
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.evaluations.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/vendors/{vendorId}/evaluations
 * Operation: startVendorEvaluation
 */
export async function startVendorEvaluation(
  request: FastifyRequest<{ Params: { vendorId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: VendorsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: StartVendorEvaluationInput = {
    vendorId: request.params.vendorId,
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.evaluations.create.execute(input);
  return reply.code(201).send(result);
}